% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{brute_force_knapsack function}
\usage{
brute_force_knapsack(x, W, parallel = FALSE)
}
\arguments{
\item{x}{as a \code{data.frame} containing the weights and values of knapsack objects with column names 'w' & 'v' respectively}

\item{W}{as a \code{numeric value} defining the constraint of the knapsack capacity}
}
\value{
list containing sum of values of chosen elements and the reference of the elements chosen in the final solution
}
\description{
evaluates the solution for knapsack problem by calculating sum of weights & profits for each combination possible for given input
}
