% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_knapsack.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{greedy_knapsack function}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{as a \code{data.frame} containing the weights and values of knapsack objects with column names 'w' & 'v' respectively}

\item{W}{as a \code{numeric value} defining the constraint of the knapsack capacity}
}
\value{
list containing sum of values of chosen elements and the reference of the elements chosen in the final solution
}
\description{
evaluates the solution for knapsack problem by choosing elements with highest value per unit weight and using complete knapsack capacity
}
