% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_dynamic.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{knapsack_dynamic function}
\usage{
knapsack_dynamic(x, W)
}
\arguments{
\item{x}{as a \code{data.frame} containing the weights and values of knapsack objects with column names 'w' & 'v' respectively}

\item{W}{as a \code{numeric value} defining the constraint of the knapsack capacity}
}
\value{
list containing sum of values of chosen elements and the reference of the elements chosen in the final solution
}
\description{
evaluates the solution for knapsack problem using dynamic programming
}
