% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_objects.R
\name{knapsack_objects}
\alias{knapsack_objects}
\title{knapsack problem data set}
\usage{
knapsack_objects(2000)
}
\arguments{
\item{n}{as integer to represent the number of objects in the knapsack problem to be generated}
}
\description{
Creates a data frame that can be used for implementing the knapsack algorithm
}
